requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-generativeai
pdfminer.six
python-dotenv